requirements.txt
# Sperm Quantification and Motility Analysis - Dependencies

# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Computer vision and image processing
opencv-python>=4.8.0
scikit-image>=0.21.0
pillow>=10.0.0

# Tracking and filtering
filterpy>=1.4.5

# Machine learning (optional, for advanced segmentation)
scikit-learn>=1.3.0
# torch>=2.0.0  # Uncomment for deep learning features
# torchvision>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Video processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Statistical analysis
statsmodels>=0.14.0
